Bye abstract class hello React hooks, test passing
