+++ packages/graph-explorer-proxy-server/node-server.js
@@ -14,7 +14,7 @@ const aws4 = require("aws4");
 dotenv.config({ path: "../graph-explorer/.env" });
 
 const proxyLogger = pino({
-  level: process.env.LOG_LEVEL,
+  level: process.env.LOG_LEVEL || "info",
   transport: {
     target: "pino-pretty",
     options: {
@@ -79,7 +79,12 @@ const errorHandler = (error, request, response, next) => {
     );
   }
 
-  const retryFetch = async (url, headers, retryDelay = 10000, refetchMaxRetries = 1) => {
+  const retryFetch = async (
+    url,
+    headers,
+    retryDelay = 10000,
+    refetchMaxRetries = 1
+  ) => {
     // remove the existing host headers, we want ensure that we are passing the DB endpoint hostname.
     delete headers["host"];
     if (headers["aws-neptune-region"]) {
@@ -95,7 +100,7 @@ const errorHandler = (error, request, response, next) => {
     for (let i = 0; i < refetchMaxRetries; i++) {
       try {
         proxyLogger.debug("Fetching: " + url.href);
-        const res = await fetch(url.href, {headers});
+        const res = await fetch(url.href, headers);
         if (!res.ok) {
           const result = await res.json();
           proxyLogger.error("!!Request failure!!");
@@ -150,8 +155,12 @@ const errorHandler = (error, request, response, next) => {
     }
   });
 
-  app.get("/openCypher", async (req, res, next) => {
+  app.post("/openCypher", async (req, res, next) => {
     try {
+      console.log(
+        "🚀 ~ file: node-server.js:164 ~ app.post ~ req.headers:",
+        req.headers["graph-db-connection-url"]
+      );
       const response = await retryFetch(
         new URL(
           `${req.headers["graph-db-connection-url"]}/openCypher?query=` +
@@ -196,42 +205,42 @@ const errorHandler = (error, request, response, next) => {
     }
   });
 
- app.get("/logger", (req, res, next) => {
-  let message;
-  let level;
+  app.get("/logger", (req, res, next) => {
+    let message;
+    let level;
 
-  try {
-    if (req.headers["level"] === undefined) {
-      throw new Error("No log level passed.");
-    } else {
-      level = req.headers["level"];
+    try {
+      if (req.headers["level"] === undefined) {
+        throw new Error("No log level passed.");
+      } else {
+        level = req.headers["level"];
+      }
+
+      if (req.headers["message"] === undefined) {
+        throw new Error("No log message passed.");
+      } else {
+        message = req.headers["message"].replaceAll("\\", "");
+      }
+
+      if (level.toLowerCase() === "error") {
+        proxyLogger.error(message);
+      } else if (level.toLowerCase() === "warn") {
+        proxyLogger.warn(message);
+      } else if (level.toLowerCase() === "info") {
+        proxyLogger.info(message);
+      } else if (level.toLowerCase() === "debug") {
+        proxyLogger.debug(message);
+      } else if (level.toLowerCase() === "trace") {
+        proxyLogger.trace(message);
+      } else {
+        throw new Error("Tried to log to an unknown level.");
+      }
+
+      res.send("Log received.");
+    } catch (error) {
+      next(error);
     }
-
-    if (req.headers["message"] === undefined) {
-      throw new Error("No log message passed.");
-    } else {
-      message = req.headers["message"].replaceAll("\\", "");
-    }
-
-    if (level.toLowerCase() === "error") {
-      proxyLogger.error(message);
-    } else if (level.toLowerCase() === "warn") {
-      proxyLogger.warn(message);
-    } else if (level.toLowerCase() === "info") {
-      proxyLogger.info(message);
-    } else if (level.toLowerCase() === "debug") {
-      proxyLogger.debug(message);
-    } else if (level.toLowerCase() === "trace") {
-      proxyLogger.trace(message);
-    } else {
-      throw new Error("Tried to log to an unknown level.");
-    }
-
-    res.send("Log received.");
-  } catch (error) {
-    next(error);
-  }
-});
+  });
 
   app.use(errorHandler);
