+++ packages/graph-explorer/src/connector/gremlin/queries/fetchNeighbors.ts
@@ -1,7 +1,7 @@
 import type {
   NeighborsRequest,
   NeighborsResponse,
-} from "../../AbstractConnector";
+} from "../../useGEFetchTypes";
 import mapApiEdge from "../mappers/mapApiEdge";
 import mapApiVertex from "../mappers/mapApiVertex";
 import toStringId from "../mappers/toStringId";
