+++ packages/graph-explorer/src/connector/gremlin/queries/fetchNeighborsCount.ts
@@ -1,7 +1,7 @@
 import type {
   NeighborsCountRequest,
   NeighborsCountResponse,
-} from "../../AbstractConnector";
+} from "../../useGEFetchTypes";
 import neighborsCountTemplate from "../templates/neighborsCountTemplate";
 import type { GInt64 } from "../types";
 import { GremlinFetch } from "../types";
