+++ packages/graph-explorer/src/connector/openCypher/templates/neighborsCountTemplate.ts
@@ -1,4 +1,4 @@
-import type { NeighborsCountRequest } from "../../AbstractConnector";
+import type { NeighborsCountRequest } from "../../useGEFetchTypes";
 
 /**
  * Given a single nodes id, it returns an OpenCypher template with
@@ -19,7 +19,7 @@ const neighborsCountTemplate = ({
   limit = 500,
 }: NeighborsCountRequest) => {
   let template = "";
-  template = `MATCH (v) -[e]- (t) WHERE ID(v) = \"${vertexId}\" RETURN labels(t) AS vertexLabel, count(DISTINCT t) AS count`;
+  template = `MATCH (v) -[e]- (t) WHERE ID(v) = "${vertexId}" RETURN labels(t) AS vertexLabel, count(DISTINCT t) AS count`;
 
   if (limit > 0) {
     template += ` LIMIT ${limit}`;
