+++ packages/graph-explorer/src/connector/openCypher/templates/oneHopTemplate.ts
@@ -1,4 +1,4 @@
-import type { Criterion, NeighborsRequest } from "../../AbstractConnector";
+import type { Criterion, NeighborsRequest } from "../../useGEFetchTypes";
 
 const criterionNumberTemplate = ({
   name,
@@ -102,43 +102,43 @@ const criterionTemplate = (criterion: Criterion): string => {
  * LIMIT 10
  */
 const oneHopTemplate = ({
-    vertexId,
-    filterByVertexTypes = [],
-    edgeTypes = [],
-    filterCriteria = [],
-    limit = 10,
-    offset = 0,
-  }: Omit<NeighborsRequest, "vertexType">): string => {
-    let template = `MATCH (v)`;
-  
-    const formattedVertexTypes = filterByVertexTypes
-      .flatMap(type => type.split("::"))
-      .map(type => `v:${type}`)
-      .join(" OR ");
-    const formattedEdgeTypes = edgeTypes.map(type => `${type}`).join("|");
+  vertexId,
+  filterByVertexTypes = [],
+  edgeTypes = [],
+  filterCriteria = [],
+  limit = 10,
+  offset = 0,
+}: Omit<NeighborsRequest, "vertexType">): string => {
+  let template = `MATCH (v)`;
 
-    if (edgeTypes.length > 0) {
-        template += `-[e:${formattedEdgeTypes}]-`;
-    } else {
-        template += `-[e]-`;
-    }
-    
-    if (filterByVertexTypes.length == 1) {
-      template += `(tgt:${filterByVertexTypes[0]}) WHERE ID(v) = \"${vertexId}\" `;
-    } else if (filterByVertexTypes.length > 1) {
-      template += `(tgt) WHERE ID(v) = \"${vertexId}\" AND ${formattedVertexTypes}`;
-    } else {
-      template += `(tgt) WHERE ID(v) = \"${vertexId}\" `;
-    }
-  
-    let filterCriteriaTemplate = filterCriteria?.map(criterionTemplate).join(" AND ");
-    if (filterCriteriaTemplate) {
-      template += `AND ${filterCriteriaTemplate} `;
-    }
+  const formattedVertexTypes = filterByVertexTypes
+    .flatMap(type => type.split("::"))
+    .map(type => `v:${type}`)
+    .join(" OR ");
+  const formattedEdgeTypes = edgeTypes.map(type => `${type}`).join("|");
 
-    template += `WITH collect(DISTINCT tgt)[..${limit}] AS vObjects, collect({edge: e, sourceType: labels(v), targetType: labels(tgt)})[..${limit}] AS eObjects RETURN vObjects, eObjects`;
-  
-    return template;
-  };
+  if (edgeTypes.length > 0) {
+    template += `-[e:${formattedEdgeTypes}]-`;
+  } else {
+    template += `-[e]-`;
+  }
+
+  if (filterByVertexTypes.length == 1) {
+    template += `(tgt:${filterByVertexTypes[0]}) WHERE ID(v) = \"${vertexId}\" `;
+  } else if (filterByVertexTypes.length > 1) {
+    template += `(tgt) WHERE ID(v) = \"${vertexId}\" AND ${formattedVertexTypes}`;
+  } else {
+    template += `(tgt) WHERE ID(v) = \"${vertexId}\" `;
+  }
+
+  const filterCriteriaTemplate = filterCriteria?.map(criterionTemplate).join(" AND ");
+  if (filterCriteriaTemplate) {
+    template += `AND ${filterCriteriaTemplate} `;
+  }
+
+  template += `WITH collect(DISTINCT tgt)[..${limit}] AS vObjects, collect({edge: e, sourceType: labels(v), targetType: labels(tgt)})[..${limit}] AS eObjects RETURN vObjects, eObjects`;
+
+  return template;
+};
 
 export default oneHopTemplate;
