+++ packages/graph-explorer/src/connector/utils/isErrorResponse.ts
@@ -1,4 +1,4 @@
-import { ErrorResponse } from "../AbstractConnector";
+import type { ErrorResponse } from "../useGEFetchTypes";
 
 const isErrorResponse = (value: any): value is ErrorResponse => {
   return !!(value.code && value.detailedMessage);
